Avoid truncation warning on Sun Workshop compiler. (JIRA: CXX-893)
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@41335 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/misc/xmlwrapp/event_parser.cpp
@@ -141,7 +141,7 @@ bool xml::event_parser::parse_stream (std::istream &stream) {
 }
 //####################################################################
 bool xml::event_parser::parse_chunk (const char *chunk, size_type length) {
-    xmlParseChunk(pimpl_->parser_context_, chunk, length, 0);
+    xmlParseChunk(pimpl_->parser_context_, chunk, static_cast<int>(length), 0);
     return pimpl_->parser_status_;
 }
 //####################################################################
